[Config]: Add 'mods_config.cfg'

+++ src/Components/Modules/Config.cpp
@@ -105,21 +105,60 @@ namespace Components
 		return {};
 	}
 
-	void Config::ReplaceConfig(char* __str, signed __int32 __size, const char* __format, ...)
+	void Config::CallExecFromCFG(bool useModConfig)
 	{
-		//original function
-		Utils::Hook::Call<int(char*, int, const char*)>(0x532A40)(__str, __size, __format); //Com_SetPlayerProfile
-
+		Command::Execute("disconnect");
+		Utils::Hook::Call<void(std::uint32_t, int)>(0x589DB0)(0xFFFF, 0);
 		const std::string& profileName = Utils::IO::ReadFile("players/profiles/active.txt");
-		const std::string& profilePath = Utils::String::VA("players/profiles/%s/", profileName.c_str());
+		Dvars::Functions::Dvar_SetStringByName("com_playerProfile", profileName.c_str());
 
-		if (Utils::IO::FileExists(profilePath.c_str() + "config.cfg"s))
+		char __str[64];
+		if (!useModConfig)
 		{
-			auto data = Utils::IO::ReadFile(profilePath.c_str() + "config.cfg"s);
-			Utils::IO::WriteFile(profilePath.c_str() + "config_backup.cfg"s, data);
-			Utils::IO::RemoveFile(profilePath.c_str() + "config.cfg"s);
-			Utils::IO::RemoveFile(profilePath.c_str() + "iw3sp_mod_config.cfg"s);
-			Utils::IO::WriteFile(profilePath.c_str() + "iw3sp_mod_config.cfg"s, data);
+			Utils::Hook::Call<void(char*, int, char const*)>(0x532A40)(__str, 64, CLIENT_CONFIG);
+		}
+		else
+		{
+			Utils::Hook::Call<void(char*, int, char const*)>(0x532A40)(__str, 64, MOD_CONFIG);
+
+			const std::string& moddir = Dvars::Functions::Dvar_FindVar("fs_game")->current.string;
+			const std::string& profilePathStock = Utils::String::VA("players/profiles/%s/", profileName.c_str());
+
+			// Reallocating settings from one .cfg to another
+			if (!Utils::IO::FileExists(profilePathStock.c_str() + "DO_NOT_DELETE"s))
+			{
+				auto data = Utils::IO::ReadFile(profilePathStock.c_str() + "iw3sp_mod_config.cfg"s);
+				Utils::IO::WriteFile(profilePathStock.c_str() + "mods_config.cfg"s, data);
+				Utils::IO::WriteFile(profilePathStock.c_str() + "DO_NOT_DELETE"s, "// DO NOT DELETE THIS FILE IF YOU WANT KEEP THE SETTINGS IN 'mods_config.cfg'!", true);
+			}
+		}
+
+		Game::Com_ExecStartupConfigs(Game::uiInfo->uiDC.localClientNum, __str);
+	}
+
+	void Config::ReplaceConfig(char* __str, signed __int32 __size, const char* __format, ...)
+	{	
+		const std::string& profileName = Utils::IO::ReadFile("players/profiles/active.txt");
+
+		if (!Config::GameWithMod)
+		{
+			Utils::Hook::Call<void(char*, int, const char*)>(0x532A40)(__str, __size, CLIENT_CONFIG); //Com_BuildPlayerProfilePath
+
+			const std::string& profilePath = Utils::String::VA("players/profiles/%s/", profileName.c_str());
+
+			if (Utils::IO::FileExists(profilePath.c_str() + "config.cfg"s))
+			{
+				// Reallocating settings from one .cfg to another
+				auto data = Utils::IO::ReadFile(profilePath.c_str() + "config.cfg"s);
+				Utils::IO::WriteFile(profilePath.c_str() + "config_backup.cfg"s, data);
+				Utils::IO::RemoveFile(profilePath.c_str() + "config.cfg"s);
+				Utils::IO::RemoveFile(profilePath.c_str() + "iw3sp_mod_config.cfg"s);
+				Utils::IO::WriteFile(profilePath.c_str() + "iw3sp_mod_config.cfg"s, data);
+			}
+		}
+		else
+		{
+			Utils::Hook::Call<void(char*, int, const char*)>(0x532A40)(__str, __size, MOD_CONFIG); //Com_BuildPlayerProfilePath
 		}
 	}
 
@@ -145,7 +184,21 @@ namespace Components
 
 	Config::Config()
 	{
-		//Com_SetPlayerProfile
+		if (!Config::GameWithMod)
+		{
+			Utils::Hook::Set<const char*>(0x5309CE, CLIENT_CONFIG);
+			Utils::Hook::Set<const char*>(0x532778, CLIENT_CONFIG);
+			Utils::Hook::Set<const char*>(0x53544A, CLIENT_CONFIG);
+			Utils::Hook::Set<const char*>(0x57B2BD, CLIENT_CONFIG);
+		}
+		else
+		{
+			Utils::Hook::Set<const char*>(0x5309CE, MOD_CONFIG);
+			Utils::Hook::Set<const char*>(0x532778, MOD_CONFIG);
+			Utils::Hook::Set<const char*>(0x53544A, MOD_CONFIG);
+			Utils::Hook::Set<const char*>(0x57B2BD, MOD_CONFIG);
+		}
+
 		Utils::Hook(0x532783, Com_SetPlayerProfile_stub, HOOK_JUMP).install()->quick();
 		Utils::Hook(0x535455, Com_SetPlayerProfile_stub2, HOOK_JUMP).install()->quick();
